Reorganized some code. Fixed level gen difficulty

+++ LudumDare34/Assets/Scripts/IEnemy.cs
@@ -3,5 +3,5 @@ using System.Collections;
 
 public interface IEnemy {
 	void FaceRight();
-	void SetDifficulty (int dif);
+	void SetDifficulty (float dif);
 }
